+++ pysubgroup/algorithms.py
@@ -4,10 +4,11 @@ Created on 29.04.2016
 @author: lemmerfn
 '''
 import copy
+from math import factorial
 from itertools import combinations, chain
 from heapq import heappush, heappop
 from collections import Counter, namedtuple
-
+import warnings
 import numpy as np
 import pysubgroup as ps
 
@@ -28,8 +29,11 @@ class SubgroupDiscoveryTask:
         self.weighting_attribute = weighting_attribute
 
 
+
+
+
 class Apriori:
-    def __init__(self, representation_type=None, combination_name='Conjunction'):
+    def __init__(self, representation_type=None, combination_name='Conjunction', use_numba=True):
         self.combination_name = combination_name
 
         if representation_type is None:
@@ -38,7 +42,15 @@ class Apriori:
         self.use_vectorization = True
         self.use_repruning = False
         self.optimistic_estimate_name = 'optimistic_estimate'
-
+        self.next_level = self.get_next_level
+        self.compiled_func = None
+        if use_numba:
+            try:
+                import numba # pylint: disable=unused-import
+                self.next_level = self.get_next_level_numba
+                print('Apriori: Using numba for speedup')
+            except ImportError:
+                pass
     def get_next_level_candidates(self, task, result, next_level_candidates):
         promising_candidates = []
         optimistic_estimate_function = getattr(task.qf, self.optimistic_estimate_name)
@@ -84,14 +96,47 @@ class Apriori:
                                         if all(frozenset(comb) not in unpromising_combinations for comb in combinations(selectors, k)))
         return promising_candidates
 
+    def get_next_level_numba(self, promising_candidates):
+
+        from numba import jit
+        if not hasattr(self, 'compiled_func'):
+            @jit
+            def getNewCandidates(l, hashes):
+                result = []
+                for i in range(len(l)-1):
+                    for j in range(i + 1, len(l)):
+                        if hashes[i] == hashes[j]:
+                            if np.all(l[i, :-1] == l[j, :-1]):
+                                result.append((i, j))
+                return result
+            self.compiled_func = getNewCandidates
+
+        all_selectors = Counter(chain.from_iterable(promising_candidates))
+        d = {selector:i for i, selector in enumerate(all_selectors)}
+        l = [tuple(d[sel] for sel in selectors) for selectors in promising_candidates]
+        arr = np.array(l, dtype=int)
+
+        print(len(arr))
+        hashes = np.array([hash(tuple(x[:-1])) for x in l], dtype=np.int64)
+        candidates_int = self.compiled_func(arr, hashes)
+        return list((*promising_candidates[i], promising_candidates[j][-1])  for i, j in candidates_int)
+
+    def get_next_level(self, promising_candidates):
+        precomputed_list = list((tuple(sg), sg[-1], hash(tuple(sg[:-1])), tuple(sg[:-1])) for sg in promising_candidates)
+        return list((*sg1, new_selector) for (sg1, _, hash_l, selectors_l), (_, new_selector, hash_r, selectors_r) in combinations(precomputed_list, 2)
+                    if (hash_l == hash_r) and (selectors_l == selectors_r))
+
+
 
     def execute(self, task):
         if not isinstance(task.qf, ps.BoundedInterestingnessMeasure):
             raise RuntimeWarning("Quality function is unbounded, long runtime expected")
+
+        task.qf.calculate_constant_statistics(task)
+        
         with self.representation_type(task.data) as representation:
             combine_selectors = getattr(representation.__class__, self.combination_name)
-            result = []
-            task.qf.calculate_constant_statistics(task)
+            result = []    
             # init the first level
             next_level_candidates = []
             for sel in task.search_space:
@@ -113,17 +158,13 @@ class Apriori:
                 if self.use_repruning:
                     promising_candidates = self.reprune_lower_levels(promising_candidates, depth)
 
-                set_promising_candidates = set(frozenset(p) for p in promising_candidates)
+                next_level_candidates_no_pruning = self.next_level(promising_candidates)
 
-                precomputed_list = list((sg, [sg[-1]], hash(tuple(sg[:-1])), sg[:-1]) for sg in promising_candidates)
-                next_level_candidates_no_pruning = (sg1 + new_selector for (sg1, _, hash_l, selectors_l), (_, new_selector, hash_r, selectors_r) in combinations(precomputed_list, 2)
-                                                    if (hash_l == hash_r) and (selectors_l == selectors_r))
-                del precomputed_list
                 # select those selectors and build a subgroup from them
                 #   for which all subsets of length depth (=candidate length -1) are in the set of promising candidates
-
+                set_promising_candidates = set(tuple(p) for p in promising_candidates)
                 next_level_candidates = [ps.Subgroup(task.target, combine_selectors(selectors)) for selectors in next_level_candidates_no_pruning
-                                         if all((frozenset(subset) in set_promising_candidates) for subset in combinations(selectors, depth))]
+                                         if all((subset in set_promising_candidates) for subset in combinations(selectors, depth))]
                 depth = depth + 1
 
         result.sort(key=lambda x: x[0], reverse=True)
@@ -133,7 +174,7 @@ class Apriori:
 class BestFirstSearch:
     def execute(self, task):
         result = []
-        queue = []
+        queue = [(float("-inf"), ps.Conjunction([]))]
         operator = ps.StaticSpecializationOperator(task.search_space)
         task.qf.calculate_constant_statistics(task)
         # init the first level
@@ -145,17 +186,15 @@ class BestFirstSearch:
             q = -q
             if not (q > ps.minimum_required_quality(result, task)):
                 break
-            sg = ps.Subgroup(task.target, candidate_description)
-            statistics = task.qf.calculate_statistics(sg, task.data)
-            ps.add_if_required(result, sg, task.qf.evaluate(sg, statistics), task)
-            optimistic_estimate = task.qf.optimistic_estimate(sg, statistics)
+            for candidate_description in operator.refinements(old_description):
+                sg = ps.Subgroup(task.target, candidate_description)
+                statistics = task.qf.calculate_statistics(sg, task.data)
+                ps.add_if_required(result, sg, task.qf.evaluate(sg, statistics), task)
+                optimistic_estimate = task.qf.optimistic_estimate(sg, statistics)
 
-
-            # compute refinements and fill the queue
-            if len(candidate_description) < task.depth and optimistic_estimate >= ps.minimum_required_quality(result, task):
-                #print(ps.minimum_required_quality(result, task))
-                for new_description in operator.refinements(candidate_description):
-                    heappush(queue, (-optimistic_estimate, new_description))
+                # compute refinements and fill the queue
+                if len(candidate_description) < task.depth and optimistic_estimate >= ps.minimum_required_quality(result, task):
+                    heappush(queue, (-optimistic_estimate, candidate_description))
 
         result.sort(key=lambda x: x[0], reverse=True)
         return result
@@ -268,10 +307,25 @@ class BeamSearch:
 
 
 class SimpleSearch:
+    def __init__(self, show_progress=True):
+        self.show_progress = show_progress
     def execute(self, task):
         task.qf.calculate_constant_statistics(task)
         result = []
         all_selectors = chain.from_iterable(combinations(task.search_space, r) for r in range(1, task.depth + 1))
+        if self.show_progress:
+            try:
+                from tqdm import tqdm
+                def binomial(x, y):
+                    try:
+                        binom = factorial(x) // factorial(y) // factorial(x - y)
+                    except ValueError:
+                        binom = 0
+                    return binom
+                total = sum(binomial(len(task.search_space), k) for k in range(1, task.depth + 1))
+                all_selectors = tqdm(all_selectors, total=total)
+            except ImportError:
+                pass
         for selectors in all_selectors:
             sg = ps.Subgroup(task.target, ps.Conjunction(selectors))
             statistics = task.qf.calculate_statistics(sg, task.data)
@@ -296,12 +350,13 @@ class SimpleDFS:
             optimistic_estimate = task.qf.optimistic_estimate(sg, statistics)
             if not(optimistic_estimate > ps.minimum_required_quality(result, task)):
                 return result
-
+        if statistics.size < 200:
+            return result
 
         quality = task.qf.evaluate(sg, statistics)
         ps.add_if_required(result, sg, quality, task)
 
-        if len(prefix) < task.depth:
+        if len(prefix) < task.depth and statistics.size > 200:
             new_modification_set = copy.copy(modification_set)
             for sel in modification_set:
                 prefix.append(sel)
@@ -324,26 +379,24 @@ class DFS:
         self.params_tpl = namedtuple('StandardQF_parameters', ('size', 'positives_count'))
 
     def execute(self, task):
-        self.target_bitset = task.target.covers(task.data)
         self.operator = ps.StaticSpecializationOperator(task.search_space)
         task.qf.calculate_constant_statistics(task)
         result = []
-        with self.apply_representation(task.data):
+        with self.apply_representation(task.data, task.search_space):
             self.search_internal(task, result, ps.RepresentationConjunction([]))
         result.sort(key=lambda x: x[0], reverse=True)
         result = [(quality, ps.Subgroup(task, sgd)) for quality, sgd in result]
         return result
 
     def search_internal(self, task, result, sg):
-        params = self.params_tpl(sg.size, np.count_nonzero(self.target_bitset[sg]))
-
-        optimistic_estimate = task.qf.optimistic_estimate(sg, params)
+        statistics = task.qf.calculate_statistics(sg)
+        optimistic_estimate = task.qf.optimistic_estimate(sg, statistics)
         if not(optimistic_estimate > ps.minimum_required_quality(result, task)):
             return
-        quality = task.qf.evaluate(sg, params)
+        quality = task.qf.evaluate(sg, statistics)
         ps.add_if_required(result, sg, quality, task)
 
-        if len(sg) < task.depth:
+        if sg.depth < task.depth and statistics.size > 0:
             for new_sg in self.operator.refinements(sg):
                 self.search_internal(task, result, new_sg)
 
@@ -355,18 +408,19 @@ class DFSNumeric():
         self.f = None
         self.target_values = None
         self.bitsets = {}
+        self.num_calls = 0
 
     def execute(self, task):
         if not isinstance(task.qf, ps.StandardQFNumeric):
-            raise NotImplementedError("BSD_numeric so far is only implemented for StandardQFNumeric")
+            warnings.warn("BSD_numeric so far is only implemented for StandardQFNumeric")
         self.pop_size = len(task.data)
         sorted_data = task.data.sort_values(task.target.get_attributes(), ascending=False)
 
         # generate target bitset
-        self.target_values = sorted_data[task.target.get_attributes()[0]].values
+        self.target_values = sorted_data[task.target.get_attributes()[0]].to_numpy()
 
         task.qf.calculate_constant_statistics(task)
-        self.f = task.qf.evaluate
+        self.evaluate = task.qf.evaluate
 
         # generate selector bitsets
         self.bitsets = {}
@@ -379,6 +433,7 @@ class DFSNumeric():
         return result
 
     def search_internal(self, task, prefix, modification_set, result, bitset):
+        self.num_calls += 1
         sg_size = bitset.sum()
         if sg_size == 0:
             return result
@@ -388,7 +443,7 @@ class DFSNumeric():
         sizes = np.arange(1, len(target_values_cs) + 1)
         mean_values_cs = target_values_cs / sizes
         tpl = DFSNumeric.tpl(sizes, mean_values_cs)
-        qualities = self.f(None, tpl)
+        qualities = self.evaluate(None, tpl)
         optimistic_estimate = np.max(qualities)
 
         if optimistic_estimate <= ps.minimum_required_quality(result, task):