+++ pysubgroup/nominal_target.py
@@ -10,7 +10,7 @@ import scipy.stats
 
 import pysubgroup as ps
 
-from .subgroup import Subgroup, NominalSelector
+from .subgroup import Subgroup, EqualitySelector
 from .boolean_expressions import Conjunction
 
 
@@ -25,7 +25,7 @@ class NominalTarget():
         if target_attribute is not None and target_value is not None:
             if target_selector is not None:
                 raise BaseException("NominalTarget is to be constructed EITHER by a selector OR by attribute/value pair")
-            target_selector = NominalSelector(target_attribute, target_value)
+            target_selector = EqualitySelector(target_attribute, target_value)
         if target_selector is None:
             raise BaseException("No target selector given")
         self.target_selector = target_selector
@@ -106,7 +106,8 @@ class NominalTarget():
         subgroup.statistics['lift_weighted'] = (positives_subgroup / instances_subgroup) / (positives_dataset / instances_dataset)
 
 
-class SimplePositivesQF(ps.AbstractInterestingnessMeasure):
+
+class SimplePositivesQF(ps.AbstractInterestingnessMeasure): # pylint: disable=abstract-method
     tpl = namedtuple('PositivesQF_parameters', ('size', 'positives_count'))
 
     def __init__(self):
@@ -124,32 +125,65 @@ class SimplePositivesQF(ps.AbstractInterestingnessMeasure):
     def calculate_statistics(self, subgroup, data=None):
         if hasattr(subgroup, "representation"):
             cover_arr = subgroup
+            size = subgroup.size
         else:
             cover_arr = subgroup.covers(data)
-        return SimplePositivesQF.tpl(np.count_nonzero(cover_arr), np.count_nonzero(self.positives[cover_arr]))
+            size = np.count_nonzero(cover_arr)
+        return SimplePositivesQF.tpl(size, np.count_nonzero(self.positives[cover_arr]))
 
     def is_applicable(self, subgroup):
         return isinstance(subgroup.target, NominalTarget)
 
-        
+
+
+# TODO Make ChiSquared useful for real nominal data not just binary
+# TODO Introduce Enum for direction
+# TODO Maybe it is possible to give a optimistic estimate for ChiSquared
 class ChiSquaredQF(SimplePositivesQF):
+    """
+    ChiSquaredQF which test for statistical independence of a subgroup against it's complement
+
+    ...
+
+    """
+
     @staticmethod
-    def chi_squared_qf(instances_dataset, positives_dataset, instances_subgroup, positives_subgroup, min_instances=5, bidirect=True, direction_positive=True):
+    def chi_squared_qf(instances_dataset, positives_dataset, instances_subgroup, positives_subgroup, min_instances=5, bidirect=True, direction_positive=True, index=0):
+        """
+        Performs chi2 test of statistical independence
+
+        Test whether a subgroup is statistically independent from it's complement (see scipy.stats.chi2_contingency).
+
+
+        Parameters
+        ----------
+        instances_dataset, positives_dataset, instances_subgroup, positives_subgroup : int
+            counts of subgroup and dataset
+        min_instances : int, optional
+            number of required instances, if less -inf is returned for that subgroup
+        bidirect : bool, optional
+            If true both directions are considered interesting else direction_positive decides which direction is interesting
+        direction_positive: bool, optional
+            Only used if bidirect=False; specifies whether you are interested in positive (True) or negative deviations
+        index : {0, 1}, optional
+            decides whether the test statistic (0) or the p-value (1) should be used
+        """
+
         if (instances_subgroup < min_instances) or ((instances_dataset - instances_subgroup) < min_instances):
             return float("-inf")
-        p_subgroup = positives_subgroup / instances_subgroup
-        p_dataset = positives_dataset / instances_dataset
-        
-        negatives_subgroup =  instances_subgroup - positives_subgroup
-        negatives_dataset =    instances_dataset - positives_dataset
+
+        negatives_subgroup =  instances_subgroup - positives_subgroup # pylint: disable=bad-whitespace
+        negatives_dataset =    instances_dataset - positives_dataset # pylint: disable=bad-whitespace
         negatives_complement = negatives_dataset - negatives_subgroup
         positives_complement = positives_dataset - positives_subgroup
 
         val = scipy.stats.chi2_contingency([[positives_subgroup, positives_complement],
-                                            [negatives_subgroup, negatives_complement]], correction=False)[0]
+                                            [negatives_subgroup, negatives_complement]], correction=False)[index]
         if bidirect:
             return val
-        elif direction_positive and p_subgroup > p_dataset:
+        p_subgroup = positives_subgroup / instances_subgroup
+        p_dataset = positives_dataset / instances_dataset
+        if direction_positive and p_subgroup > p_dataset:
             return val
         elif not direction_positive and p_subgroup < p_dataset:
             return val
@@ -173,8 +207,19 @@ class ChiSquaredQF(SimplePositivesQF):
                                             [negatives_subgroup, negatives_complement]], correction=True)[0]
         return scipy.stats.chi2.sf(val * effective_sample_size / instancesDataset, 1)
 
-    def __init__(self, direction='bidirect', min_instances=5):
-        if direction == 'bidirect':
+    def __init__(self, direction='both', min_instances=5, stat='chi2'):
+        """
+        Parameters
+        ----------
+        direction : {'both', 'positive', 'negative'}
+            direction of deviation that is of interest
+        min_instances : int, optional
+            number of required instances, if less -inf is returned for that subgroup
+        stat : {'chi2', 'p'}
+            whether to report the test statistic or the p-value (see scipy.stats.chi2_contingency)
+        """
+
+        if direction == 'both':
             self.bidirect = True
             self.direction_positive = True
         if direction == 'positive':
@@ -184,25 +229,14 @@ class ChiSquaredQF(SimplePositivesQF):
             self.bidirect = False
             self.direction_positive = False
         self.min_instances = min_instances
+        self.index = {'chi2' : 0, 'p': 1}[stat]
         super().__init__()
 
-
-    # def evaluate_from_dataset(self, data, subgroup, weighting_attribute=None):
-    #     if not self.is_applicable(subgroup):
-    #         raise BaseException("Quality measure cannot be used for this target class")
-    #     if weighting_attribute is None:
-    #         result = self.evaluate_from_statistics(*subgroup.get_base_statistics(data))
-    #     else:
-    #         (instancesDataset, positivesDataset, instancesSubgroup, positivesSubgroup) = subgroup.get_base_statistics(data, weighting_attribute)
-    #         weights = data[weighting_attribute]
-    #         base = self.evaluate_from_statistics(instancesDataset, positivesDataset, instancesSubgroup, positivesSubgroup)
-    #         result = base * ps.effective_sample_size(weights) / instancesDataset
-    #     return result
 
     def evaluate(self, subgroup, statistics=None):
         statistics = self.ensure_statistics(subgroup, statistics)
         datatset = self.datatset
-        return ChiSquaredQF.chi_squared_qf(datatset.size, datatset.positives_count, statistics.size, statistics.positives_count, self.min_instances, self.bidirect, self.direction_positive)
+        return ChiSquaredQF.chi_squared_qf(datatset.size, datatset.positives_count, statistics.size, statistics.positives_count, self.min_instances, self.bidirect, self.direction_positive, self.index)
 
     def supports_weights(self):
         return True
@@ -211,6 +245,18 @@ class ChiSquaredQF(SimplePositivesQF):
 
 
 class StandardQF(SimplePositivesQF, ps.BoundedInterestingnessMeasure):
+    """
+    StandardQF which weights the relative size against the difference in averages
+
+    The StandardQF is a general form of quality function which for different values of a is order equivalen to
+    many popular quality measures.
+
+    Attributes
+    ----------
+    a : float
+        used as an exponent to scale the relative size to the difference in averages
+
+    """
 
     @staticmethod
     def standard_qf(a, instances_dataset, positives_dataset, instances_subgroup, positives_subgroup):
@@ -222,18 +268,25 @@ class StandardQF(SimplePositivesQF, ps.BoundedInterestingnessMeasure):
         return (instances_subgroup / instances_dataset) ** a * (p_subgroup - p_dataset)
 
     def __init__(self, a):
+        """
+        Parameters
+        ----------
+        a : float
+            exponent to scale the relative size to the difference in means
+        """
+
         self.a = a
         super().__init__()
-        
+
     def evaluate(self, subgroup, statistics=None):
         statistics = self.ensure_statistics(subgroup, statistics)
         datatset = self.datatset
-        return StandardQF.standard_qf(self.a, datatset.size, datatset.positives_count, statistics.size, statistics.positives_count )
+        return StandardQF.standard_qf(self.a, datatset.size, datatset.positives_count, statistics.size, statistics.positives_count)
 
     def optimistic_estimate(self, subgroup, statistics=None):
         statistics = self.ensure_statistics(subgroup, statistics)
         datatset = self.datatset
-        return StandardQF.standard_qf(self.a, datatset.size, datatset.positives_count, statistics.positives_count, statistics.positives_count )
+        return StandardQF.standard_qf(self.a, datatset.size, datatset.positives_count, statistics.positives_count, statistics.positives_count)
 
     def optimistic_generalisation(self, subgroup, statistics=None):
         statistics = self.ensure_statistics(subgroup, statistics)
@@ -245,24 +298,58 @@ class StandardQF(SimplePositivesQF, ps.BoundedInterestingnessMeasure):
         return True
 
 
-class WRAccQF(StandardQF):
-    def __init__(self, a):
-        super().__init__(a)
-        self.a = 1.0
 
 
 class LiftQF(StandardQF):
-    def __init__(self, a):
-        super().__init__(a)
-        self.a = 0.0
+    """
+    Lift Quality Function
+
+    LiftQF is a StandardQF with a=0.
+    Thus it treats the difference in ratios as the quality without caring about the relative size of a subgroup.
+    """
+
+    def __init__(self):
+        """
+        """
+
+        super().__init__(0.0)
 
 
-class SimpleBinomial(StandardQF):
-    def __init__(self, a):
-        super().__init__(a)
-        self.a = 0.5
+
+# TODO add true binomial quality function as in https://opus.bibliothek.uni-wuerzburg.de/opus4-wuerzburg/frontdoor/index/index/docId/1786
+class SimpleBinomialQF(StandardQF):
+    """
+    Simple Binomial Quality Function
+
+    SimpleBinomialQF is a StandardQF with a=0.5.
+    It is an order equivalent approximation of the full binomial test if the subgroup size is much smaller than the size of the entire dataset.
+    """
+
+    def __init__(self):
+        """
+        """
+
+        super().__init__(0.5)
 
 
+
+class WRAccQF(StandardQF):
+    """
+    Weighted Relative Accuracy Quality Function
+
+    WRAccQF is a StandardQF with a=1.
+    It is order equivalent to the difference in the observed and expected number of positive instances.
+    """
+
+    def __init__(self):
+        """
+        """
+
+        super().__init__(1.0)
+
+
+
+# TODO Redo Generalization Aware SGD
 #####
 # GeneralizationAware Interestingness Measures
 #####